Redirect Uri Mismatch error alignment with specification (#108)
Currently, the RedirectUriMismatch error is using a non-standard errorType. Google popularized this error, but it is not part of the standard specification. Section 4.1.2.1 specifically states
` If the request fails due to a missing, invalid, or mismatching redirection URI …`
and the valid error types listed that can be returned does not include `redirect_uri_mismatch`.

I propose of the errors listed, the one that makes the most sense is `invalid_request` additionally, we can use the `error_description` field which is optional to provide the existing message of `redirect_uri_mismatch` so the developer knows what he did wrong.